api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watchtower-api"
version = "0.1.0"
description = "Watchtower backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic-settings>=2.6",
  "sqlalchemy>=2.0",
  "alembic>=1.14",
  "psycopg2-binary>=2.9",
  "httpx>=0.28",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "google-auth>=2.0",
  "requests>=2.28",
  "croniter>=5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["watchtower_api*"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["watchtower_api", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
